:wrench: :neckbeard: Renamed extralifedonations.py to particpant.py
This is to be more in line with both readability and also pep8, etc. It holds the participant class, so it should be named participant. Deals with issuse #101
